+++ src/jawl_voicecompanion/__init__.py
@@ -3,6 +3,7 @@
 from .gateway import TextGateway
 from .hostos_policy import HostOSPolicy
 from .hostos_tools import HostOSExecutor, ToolSpec
+from .browser_adapter import BrowserAdapter
 from .models import AccessLevel, RiskClass, ToolRequest
 from .arbiter import TurnArbiter, TurnPriority
 
@@ -10,6 +11,7 @@ __all__ = [
     "AccessLevel",
     "HostOSPolicy",
     "HostOSExecutor",
+    "BrowserAdapter",
     "RiskClass",
     "TextGateway",
     "ToolRequest",